Fixed for float

+++ inc/SBNcovariance.h
@@ -104,7 +104,7 @@ namespace sbn{
 
     std::vector<int> nentries;
     std::vector<TBranch*>* branch_weight;
-    std::vector<std::map<std::string, std::vector<double> >* > f_weights;
+    std::vector<std::map<std::string, std::vector<float> >* > f_weights;
 
     std::vector<std::vector<int> > vars_i;
     std::vector<std::vector<double> > vars_d;
@@ -119,7 +119,7 @@ namespace sbn{
     std::vector<std::vector<double> > multi_vecspec2D;
     std::vector<double> vecspec2DCV;
     TStopwatch watch;
-    void ProcessEvent(const std::map<std::string, std::vector<double> >& thisfWeight,
+    void ProcessEvent(const std::map<std::string, std::vector<float> >& thisfWeight,
 		      size_t fileid,
 		      int entryid);
